+++ components/Common/Search/States/WithError.tsx
@@ -3,12 +3,12 @@ import type { FC } from 'react';
 
 import styles from './index.module.css';
 
-export const EmptyState: FC = () => {
+export const WithError: FC = () => {
   const t = useTranslations();
 
   return (
-    <div className={styles.emptyStateContainer}>
-      {t('components.search.emptyState.text')}
+    <div className={styles.searchErrorContainer}>
+      {t('components.search.searchError.text')}
     </div>
   );
 };
