+++ components/Common/Search/index.tsx
@@ -4,7 +4,7 @@ import { MagnifyingGlassIcon } from '@heroicons/react/24/outline';
 import { useTranslations } from 'next-intl';
 import { useState, type FC } from 'react';
 
-import { SearchBox } from '@/components/SearchBox/components/SearchBox';
+import { WithSearchBox } from '@/components/Common/Search/States/WithSearchBox';
 import { useKeyboardCommands } from '@/hooks/react-client';
 
 import styles from './index.module.css';
@@ -43,7 +43,7 @@ export const SearchButton: FC = () => {
         <MagnifyingGlassIcon className={styles.magnifyingGlassIcon} />
         {t('components.search.searchBox.placeholder')}
       </button>
-      {isOpen ? <SearchBox onClose={closeSearchBox} /> : null}
+      {isOpen ? <WithSearchBox onClose={closeSearchBox} /> : null}
     </>
   );
 };
